Lines starting         presentation.save(self.rendered_file)         self.last_render = Timestamp.now()
    def add_slide(self, slide: Slide) -> None:
        self.slides.append(slide)

    def save(self):
        last_edits = [ slide.last_edit for slide in self.slides ]
        last_edits.append(self.last_edit)
