+++ python-app/main.py
@@ -1,13 +1,16 @@
 from routes.hello_world import hello_world
-from routes.start_report import start_report
-from routes.edit_slide import edit_slide
-from routes.change_slide_data import change_slide_data
-from routes.render_report import render_report
-from routes.export_report import export_report
-from routes.recent_reports import recent_reports
-from routes.get_report import get_report
-from routes.import_report import import_report
-from routes.start_pivot_table import start_pivot_table
+
+from routes.slide.edit_slide import edit_slide
+from routes.slide.change_slide_data import change_slide_data
+
+from routes.report.start_report import start_report
+from routes.report.render_report import render_report
+from routes.report.export_report import export_report
+from routes.report.recent_reports import recent_reports
+from routes.report.get_report import get_report
+from routes.report.import_report import import_report
+
+import routes.pivot_table as pivot_table
 
 from flask import Flask, request, jsonify
 
@@ -25,7 +28,8 @@ export_report(app)
 recent_reports(app)
 get_report(app)
 import_report(app)
-start_pivot_table(app)
+
+app.register_blueprint(pivot_table.blueprint)
 
 if __name__ == '__main__':
     logging.basicConfig(filename='logs.log')
