+++ python-app/models/pivot_table/aggregate_function_type.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class AggregateFunctionType(Enum):
+    COUNT = 0
+    AVERAGE = 1
+
