+++ python-app/routes/pivot_table/__init__.py
@@ -1,0 +1,6 @@
+from routes.pivot_table.create_pivot_table import create_pivot_table
+from flask import Blueprint
+
+blueprint = Blueprint("pivot_table", __name__, url_prefix="/pivot_table")
+
+create_pivot_table(blueprint)
