+++ python-app/routes/report/get_report.py
@@ -1,7 +1,7 @@
 from lib.with_app_decorator import with_app
 from lib.get_or_panic import get_or_panic
 
-from models.report import Report
+from models.report.self import Report
 from models.frontend.report_record import ReportRecord
 
 from flask import request
