+++ python-app/routes/report/render_report.py
@@ -2,7 +2,7 @@ from lib.with_app_decorator import with_app
 from lib.get_or_panic import get_or_panic
 from lib.random_message import random_message, RandomMessageType
 
-from models.report import Report
+from models.report.self import Report
 
 from flask import request
