+++ python-app/routes/slide/edit_slide.py
@@ -1,7 +1,7 @@
 from lib.with_app_decorator import with_app
 from lib.get_or_panic import get_or_panic
 
-from models.report import Report
+from models.report.self import Report
 
 from flask import request
 from uuid import uuid4
